Move S3 client class to sitemap module
